refactor(runner): tighten the XmlTest snapshot fixtures and javadoc
Cover the shape GITHUB-1994 was reported with -- native injection mixed into an
@Optional parameter list -- rather than a third plain injection point, and fold
the collector listener into the test that uses it, matching the nested-listener
style of Github1649Test and FactoryInstanceTest. The list it fills is no longer
wrapped in synchronizedList: the sample suite is single-threaded, and the getter
handed the live list to an unsynchronized iteration anyway, so the wrapper
advertised a guarantee it did not provide.

Trim the argumentative tail of the LegacyParameterSnapshotter javadoc down to
the fact it was resting on, rename the private overload to snapshotParameter so
resolution between Object[] and Object is not a question, and correct the unit
test's javadoc, which claimed to use only public API while calling an internal
factory.
