+++ testng-core/src/test/java/test/inject/issue1994/IssueTest.java
@@ -2,6 +2,11 @@ package test.inject.issue1994;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.ArrayList;
+import java.util.List;
+import org.testng.IConfigurationListener;
+import org.testng.ITestListener;
+import org.testng.ITestResult;
 import org.testng.TestNG;
 import org.testng.annotations.Test;
 import org.testng.xml.XmlSuite;
@@ -15,21 +20,42 @@ public class IssueTest extends SimpleBaseTest {
     XmlSuite suite = createXmlSuite("issue1994-suite");
     XmlTest xmlTest = createXmlTest(suite, "issue1994-test", XmlTestInjectionSample.class);
 
-    InjectedXmlTestCollector collector = new InjectedXmlTestCollector();
+    Collector collector = new Collector();
     TestNG tng = create(suite);
     tng.addListener(collector);
     tng.run();
 
-    // XmlTest#clone() builds its copy with new XmlTest(suite), and that constructor registers the
-    // copy in the suite. Snapshotting the injected parameter therefore used to append a phantom
-    // <test> per invocation that received one.
     assertThat(suite.getTests()).hasSize(1);
     assertThat(suite.getTests().get(0)).isSameAs(xmlTest);
 
     // isSameAs, not containsExactly: XmlTest overrides equals, so a clone compares equal to its
     // original and equality-based assertions would not discriminate.
-    assertThat(collector.getReported())
+    assertThat(collector.reported)
         .hasSize(3)
         .allSatisfy(reported -> assertThat(reported).isSameAs(xmlTest));
+  }
+
+  /** Collects the {@link XmlTest} a result reports, which is the view a reporter would get. */
+  private static class Collector implements ITestListener, IConfigurationListener {
+
+    private final List<XmlTest> reported = new ArrayList<>();
+
+    @Override
+    public void onTestSuccess(ITestResult result) {
+      record(result);
+    }
+
+    @Override
+    public void onConfigurationSuccess(ITestResult tr) {
+      record(tr);
+    }
+
+    private void record(ITestResult result) {
+      for (Object parameter : result.getParameters()) {
+        if (parameter instanceof XmlTest) {
+          reported.add((XmlTest) parameter);
+        }
+      }
+    }
   }
 }
