+++ testng-core/src/test/java/test/inject/issue1994/XmlTestInjectionSample.java
@@ -2,12 +2,15 @@ package test.inject.issue1994;
 
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
+import org.testng.annotations.Optional;
 import org.testng.annotations.Test;
 import org.testng.xml.XmlTest;
 
 /**
  * Injects the native {@link XmlTest} into both a configuration method and a test method, so the two
- * distinct paths that feed {@code ITestResult.setParameters} are both exercised.
+ * paths that feed {@code ITestResult.setParameters} are both exercised. The reporter of GITHUB-1994
+ * hit this with the injection mixed into an {@code @Optional} parameter list, so one method keeps
+ * that shape.
  */
 public class XmlTestInjectionSample {
 
@@ -18,5 +21,5 @@ public class XmlTestInjectionSample {
   public void testMethod(XmlTest xmlTest) {}
 
   @AfterTest
-  public void afterTest(XmlTest xmlTest) {}
+  public void afterTest(XmlTest xmlTest, @Optional("chrome") String browser) {}
 }
